Update 'Sites Build Using Middleman' with links and descriptions
Adding links and descriptions should make it a little nicer for people trying
to get a sense of what's possible with Middleman. Ideally these would be a
nice gallery of thumbnails, but I'm currently too lazy for that. Apologies if
I've stuffed up the descriptions or links, please feel free to fix them up.