Move redirect handling to Netlify (#672)
Since Middleman is simply a static site generator, it's limited in how
it can handle redirects. The meta refresh tags it outputs are not
accessible or SEO-friendly, and they don't send 301 redirect headers.

Netlify, which this website is hosted on, allows us to configure
redirects server-side, which resolves all of these problems.

https://www.netlify.com/docs/redirects/

Note that Netlify needs a `_redirects` file at the root of the site, but
because Middleman ignores files that begin with an underscore, a
`_redirects` file would not be output into the built site. To get around
this, a proxy is used to rename the file from `redirects` to
`_redirects` on build.