Specify where to apply redirects
The existing documentation is unclear on what to do with the example redirect. Users might interpret it as something to run from the command line, or to be added at the top of the file to be redirected. This small change clarifies that redirects must be specified in `config.rb`